# swagger.yaml
openapi: 3.0.0
info:
  title: kVAssetTracker API
  version: 1.0.0
  description: Transformer Asset Management Platform — MVP API Contract
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://api.kVAssetTracker.com/api
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      description: Super Admin only. Returns sanitized user records with pagination.
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: search, in: query, schema: {type: string}}
        - {name: role, in: query, schema: {type: string, enum: [Super Admin, Territory Manager, Engineer, Field Technician, Viewer]}}
        - {name: territory_id, in: query, schema: {type: string}}
        - {name: service_area_id, in: query, schema: {type: string}}
        - {name: is_active, in: query, schema: {type: boolean}}
      responses:
        '200':
          $ref: '#/components/responses/UserListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Super Admin role required
    post:
      tags: [Users]
      summary: Create user
      description: Super Admin only. Password is hashed by the User model and never returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          $ref: '#/components/responses/UserResponseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Super Admin role required
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Super Admin only. Returns a sanitized user record.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/UserResponseEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Super Admin role required
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Update safe user fields
      description: Super Admin only. Does not update password, role, or activation status.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          $ref: '#/components/responses/UserResponseEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Super Admin role required
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List users through the Admin route
      description: Super Admin only alias over the tested User Management list behavior. Returns sanitized users.
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: search, in: query, schema: {type: string}}
        - {name: role, in: query, schema: {type: string, enum: [Super Admin, Territory Manager, Engineer, Field Technician, Viewer]}}
        - {name: territory_id, in: query, schema: {type: string}}
        - {name: service_area_id, in: query, schema: {type: string}}
        - {name: is_active, in: query, schema: {type: boolean}}
      responses:
        '200':
          $ref: '#/components/responses/UserListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Super Admin role required
...
  /admin/maintenance:
    get:
      tags: [Admin]
      summary: Get maintenance mode status
      description: Super Admin only. Returns the durable maintenance mode state.
      responses:
        '200':
          $ref: '#/components/responses/MaintenanceModeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Super Admin role required
    post:
      tags: [Admin]
      summary: Enable or disable maintenance mode
      description: Super Admin only. Persists maintenance state and blocks non-admin unsafe writes while enabled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceModeUpdate'
      responses:
        '200':
          $ref: '#/components/responses/MaintenanceModeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Super Admin role required

  # ─── TRANSFORMERS ─────────────────────────────────────────────────────────────
  /transformers:
    get:
      tags: [Transformers]
      summary: List all transformers
      parameters:
        - {name: territory_id, in: query, schema: {type: string}}
        - {name: service_area_id, in: query, schema: {type: string}}
        - {name: network_voltage_kv, in: query, schema: {type: integer, enum: [11, 33]}}
        - {name: operational_status, in: query, schema: {type: string, enum: [Active, Faulty, 'Under Maintenance', Decommissioned]}}
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20}}
      responses:
        '200':
          $ref: '#/components/responses/PaginatedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Transformers]
      summary: Create a new transformer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransformerCreate'
      responses:
        '201':
          $ref: '#/components/responses/ItemResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /transformers/{id}:
    get:
      tags: [Transformers]
      summary: Get transformer by ID
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/ItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Transformers]
      summary: Update transformer (Territory Manager or above)
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransformerUpdate'
      responses:
        '200':
          $ref: '#/components/responses/ItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Transformers]
      summary: Soft-delete transformer (Super Admin only)
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Transformer deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /inspections:
    get:
      tags: [Inspections]
      summary: List inspections
      parameters:
        - {name: transformer_id, in: query, schema: {type: string}}
        - {name: inspected_by, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20}}
      responses:
        '200':
          $ref: '#/components/responses/PaginatedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Inspections]
      summary: Create inspection (Engineer or above)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionCreate'
      responses:
        '201':
          $ref: '#/components/responses/ItemResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /inspections/{id}:
    get:
      tags: [Inspections]
      summary: Get inspection by ID
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/ItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Inspections]
      summary: Update inspection (Engineer or above)
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionCreate'
      responses:
        '200':
          $ref: '#/components/responses/ItemResponse'
    delete:
      tags: [Inspections]
      summary: Delete inspection (Super Admin only)
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Inspection deleted

  # ─── FAULTS ───────────────────────────────────────────────────────────────────
  /faults:
    get:
      tags: [Faults]
      summary: List faults
      parameters:
        - {name: fault_status, in: query, schema: {type: string, enum: [Reported, Acknowledged, 'In Progress', Resolved, Closed]}}
        - {name: severity, in: query, schema: {type: string, enum: [Minor, Major, Critical, 'Complete Outage']}}
        - {name: transformer_id, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20}}
      responses:
        '200':
          $ref: '#/components/responses/PaginatedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Faults]
      summary: Report a new fault (Field Technician or above)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaultCreate'
      responses:
        '201':
          $ref: '#/components/responses/ItemResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /faults/{id}:
    get:
      tags: [Faults]
      summary: Get fault by ID
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/ItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Faults]
      summary: Hard-delete fault (Super Admin only)
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Fault deleted
...
  /maintenance:
    get:
      tags: [Maintenance]
      summary: List maintenance records
      parameters:
        - {name: transformer_id, in: query, schema: {type: string}}
        - {name: maintenance_type, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20}}
      responses:
        '200':
          $ref: '#/components/responses/PaginatedResponse'
    post:
      tags: [Maintenance]
      summary: Create maintenance record (Field Technician or above)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transformer_id, maintenance_type, maintenance_date]
              properties:
                transformer_id:
                  type: string
                maintenance_type:
                  type: string
                  enum: ['Routine Service', 'Oil Change', 'Bushing Replacement', 'Winding Repair', 'Complete Overhaul', 'Emergency Repair', 'Preventive', 'Other']
                maintenance_date:
                  type: string
                  format: date
                technician_name:
                  type: string
                cost:
                  type: number
                notes:
                  type: string
      responses:
        '201':
          $ref: '#/components/responses/ItemResponse'
...
  /maintenance/{id}:
    get:
      tags: [Maintenance]
      summary: Get maintenance record by ID
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/ItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Maintenance]
      summary: Update maintenance record (Engineer or above)
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          $ref: '#/components/responses/ItemResponse'
    delete:
      tags: [Maintenance]
      summary: Delete maintenance record (Super Admin only)
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Record deleted

  # ─── TERRITORIES ──────────────────────────────────────────────────────────────
...
  /reports/transformers:
    get:
      tags: [Reports]
      summary: Generate transformer report
      parameters:
        - {name: startDate, in: query, schema: {type: string, format: date}}
        - {name: endDate, in: query, schema: {type: string, format: date}}
        - {name: territory_id, in: query, schema: {type: string}}
        - {name: service_area_id, in: query, schema: {type: string}}
        - {name: feeder_id, in: query, schema: {type: string}}
        - {name: district_id, in: query, schema: {type: string}}
        - {name: network_voltage_kv, in: query, schema: {type: number, enum: [11, 33]}}
        - {name: kva_rating, in: query, schema: {type: number, enum: [50, 100, 160, 200, 250, 315, 500, 630, 1000]}}
        - {name: operational_status, in: query, schema: {type: string, enum: [Active, Inactive, Decommissioned, 'Under Maintenance', Faulty]}}
        - {name: format, in: query, schema: {type: string, enum: [json], default: json}}
      responses:
        '200':
          $ref: '#/components/responses/ReportResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /reports/inspections:
    get:
      tags: [Reports]
      summary: Generate inspection report
      parameters:
        - {name: startDate, in: query, schema: {type: string, format: date}}
        - {name: endDate, in: query, schema: {type: string, format: date}}
        - {name: territory_id, in: query, schema: {type: string}}
        - {name: service_area_id, in: query, schema: {type: string}}
        - {name: feeder_id, in: query, schema: {type: string}}
        - {name: district_id, in: query, schema: {type: string}}
        - {name: transformer_id, in: query, schema: {type: string}}
        - {name: condition, in: query, schema: {type: string, enum: [Good, Fair, Poor, Critical]}}
        - {name: format, in: query, schema: {type: string, enum: [json], default: json}}
      responses:
        '200':
          $ref: '#/components/responses/ReportResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /reports/faults:
    get:
      tags: [Reports]
      summary: Generate fault report
      parameters:
        - {name: startDate, in: query, schema: {type: string, format: date}}
        - {name: endDate, in: query, schema: {type: string, format: date}}
        - {name: territory_id, in: query, schema: {type: string}}
        - {name: service_area_id, in: query, schema: {type: string}}
        - {name: feeder_id, in: query, schema: {type: string}}
        - {name: district_id, in: query, schema: {type: string}}
        - {name: fault_status, in: query, schema: {type: string, enum: [Open, Assigned, 'In Progress', Resolved, Closed]}}
        - {name: severity, in: query, schema: {type: string, enum: [Minor, Major, Critical, 'Complete Outage']}}
        - {name: fault_type, in: query, schema: {type: string, enum: [Overload, 'Oil Leak', 'Bushing Failure', 'Winding Failure', 'Complete Failure', Fire, Theft, Vandalism, 'LV Side Fault', 'HV Side Fault', Other]}}
        - {name: format, in: query, schema: {type: string, enum: [json], default: json}}
      responses:
        '200':
          $ref: '#/components/responses/ReportResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /reports/maintenance:
    get:
      tags: [Reports]
      summary: Generate maintenance report
      parameters:
        - {name: startDate, in: query, schema: {type: string, format: date}}
        - {name: endDate, in: query, schema: {type: string, format: date}}
        - {name: territory_id, in: query, schema: {type: string}}
        - {name: service_area_id, in: query, schema: {type: string}}
        - {name: feeder_id, in: query, schema: {type: string}}
        - {name: district_id, in: query, schema: {type: string}}
        - {name: maintenance_type, in: query, schema: {type: string, enum: [Preventive, Corrective, Emergency]}}
        - {name: format, in: query, schema: {type: string, enum: [json], default: json}}
      responses:
        '200':
          $ref: '#/components/responses/ReportResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId
  schemas:
...
    UserCreate:
      type: object
      required: [name, email, password, confirmPassword, role]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: Must include uppercase, lowercase, number, and special character.
        confirmPassword:
          type: string
          minLength: 8
        role:
          type: string
          enum: [Super Admin, Territory Manager, Engineer, Field Technician, Viewer]
        territory_id:
          type: string
          description: Required for Territory Manager, Engineer, and Field Technician.
        service_area_id:
          type: string
          description: Required for Engineer and Field Technician.
        is_active:
          type: boolean
          default: true
    UserUpdate:
      type: object
      description: Safe profile/admin-managed fields only. Password, role, and activation status are ignored by this endpoint.
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        territory_id:
          type: string
        service_area_id:
          type: string
        preferences:
          type: object
          properties:
            theme:
              type: string
              enum: [light, dark, system]
            notifications:
              type: object
              properties:
                email:
                  type: boolean
                push:
                  type: boolean
                sms:
                  type: boolean
            dashboard_widgets:
              type: array
              items:
                type: string
...
    MaintenanceModeUpdate:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        message:
          type: string
          maxLength: 500
          example: System is under maintenance
        reason:
          type: string
          maxLength: 500
          example: Planned maintenance
...
  responses:
    PaginatedResponse:
      description: Paginated list
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              message:
                type: string
              data:
                type: array
                items:
                  type: object
              pagination:
                type: object
                properties:
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  pages:
                    type: integer
...
    UserListResponse:
      description: Users retrieved successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: array
                items:
                  $ref: '#/components/schemas/User'
              pagination:
                type: object
                properties:
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  pages:
                    type: integer
    UserResponseEnvelope:
      description: User operation completed successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
              data:
                $ref: '#/components/schemas/User'
...
    MaintenanceModeResponse:
      description: Maintenance mode status returned successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
              data:
                $ref: '#/components/schemas/MaintenanceMode'
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Validation failed
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: No token provided
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Resource not found
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT